+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
  - `Environment::getInlineRendererForClass()` was replaced with `Environment::getInlineRenderersForClass()` (note the added `s`)
  - The `Environment::get____()` methods now return an iterator instead of an array
  - `Context::addBlock()` no longer returns the same block instance you passed into the method, as this served no useful purpose
+ - `RegexHelper::isEscapable()` no longer accepts `null` values
 
 ### Removed